pyproject.toml
[project]
name = "imdb"
version = "0.1.0"
description = "imdb with aiohttp server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.12",
    "aiohttp-swagger>=1.0.16",
    "aiopg[sa]>=1.4.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.1.0",
]
